USA/NORTH_CAROLINA/requirements.txt
requests==2.31.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
urllib3==1.26
boto3==1.34.63